Refactor error handling
Create custom exception
Put error throwing primarily in the repo
Split sessioncontroller and baseview into their own files.
