+++ Console/CodingTracker/CodingTracker/Views/BaseView.cs
@@ -1,0 +1,15 @@
+namespace CodingTracker.Views;
+
+public abstract class BaseView
+{
+    public void ShowSuccess(string message)
+    {
+        AnsiConsole.MarkupLine($"[green]{message}[/]");
+        Console.ReadKey();
+    }
+    public void ShowError(string message)
+    {
+        AnsiConsole.MarkupLine($"[red]{message}[/]");
+        Console.ReadKey();
+    }
+}
